+++ flycomp/src/lib.rs
@@ -5,7 +5,13 @@
 //! and dispatches to the appropriate sub-parser.
 use anyhow::Context;
 
-pub mod man;
+mod parse_help;
+pub mod parse_man;
+
+pub use parse_help::{parse_help, parse_help_argparse, parse_help_clap, parse_help_generic};
+
+#[cfg(test)]
+pub(crate) use parse_help::parse_flag_tokens;
 
 #[derive(Clone, Debug, PartialEq, Eq, Default)]
 pub enum SynthesisStrategy {
@@ -47,565 +53,6 @@ pub struct Command {
     pub args: Vec<Arg>,
     /// Recognised sub-commands.
     pub subcommands: Vec<Command>,
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Format detection
-// ──────────────────────────────────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HelpFormat {
-    /// Output generated by [clap](https://docs.rs/clap) (Rust).
-    Clap,
-    /// Output generated by Python's `argparse` module.
-    Argparse,
-    /// Unknown / generic format.
-    Generic,
-}
-
-fn detect_format(help: &str) -> HelpFormat {
-    // clap outputs a "Usage:" line (capital U) and uses long option descriptions
-    // indented underneath the flag, separated by an empty line.
-    let has_usage = help.contains("Usage:");
-    let has_commands_section = help.contains("\nCommands:\n") || help.contains("\nSubcommands:\n");
-    // argparse always prints "usage:" (lowercase) and "optional arguments:" or
-    // "options:" sections.
-    let has_positional = help.contains("positional arguments:");
-    let has_optional = help.contains("optional arguments:") || help.contains("options:");
-
-    if has_positional || (has_optional && !has_commands_section) {
-        // Pure argparse output never has a "Commands:" section.
-        HelpFormat::Argparse
-    } else if has_usage || has_commands_section {
-        HelpFormat::Clap
-    } else {
-        HelpFormat::Generic
-    }
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Public entry point
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Parse a `--help` string into a [`Command`] with the best fidelity possible.
-///
-/// The function auto-detects whether the string was produced by *clap*,
-/// Python *argparse*, or an unknown tool, and dispatches accordingly.
-pub fn parse_help(help: &str) -> Command {
-    match detect_format(help) {
-        HelpFormat::Clap => parse_help_clap(help),
-        HelpFormat::Argparse => parse_help_argparse(help),
-        HelpFormat::Generic => parse_help_generic(help),
-    }
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Helper utilities
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Leading-space count for a line (tabs count as 1).
-fn indent_of(line: &str) -> usize {
-    line.chars().take_while(|c| *c == ' ' || *c == '\t').count()
-}
-
-/// Try to parse a flag token like `-v`, `--verbose`, or `--output <FILE>`.
-/// Returns `(short, long, value_type)`.
-fn parse_flag_tokens(token: &str) -> (Option<String>, Option<String>, Option<String>) {
-    // Split on commas / whitespace to handle "-v, --verbose <FILE>"
-    let mut short = None;
-    let mut long = None;
-    let mut value_type = None;
-
-    // Tokenise: split on ", " first, then whitespace
-    let pieces: Vec<&str> = token.split_whitespace().collect();
-    for piece in &pieces {
-        if piece.starts_with("--") {
-            // May be "--flag" or "--flag=<VAL>"
-            if let Some((flag, val)) = piece.split_once('=') {
-                long = Some(flag.trim_end_matches(',').to_string());
-                value_type = Some(val.trim_matches(|c| c == '<' || c == '>').to_string());
-            } else {
-                long = Some(piece.trim_end_matches(',').to_string());
-            }
-        } else if let Some(short_candidate) = piece.strip_prefix('-') {
-            let short_candidate = short_candidate.trim_end_matches(',');
-            // Only treat single-character forms like `-v` as clap short flags.
-            // Multi-character forms such as `-wk`, `-wK`, or `-U[dlexhi]` are
-            // command-specific syntax, not plain one-letter short options.
-            if short_candidate.chars().count() == 1 {
-                short = Some(format!("-{short_candidate}"));
-            }
-        } else if piece.starts_with('<') || piece.starts_with('[') {
-            // Meta-variable — only capture the first one found so that description
-            // text like `[default: 10]` does not overwrite an already-parsed hint.
-            // Subsequent bracketed tokens (e.g. `[boolean]`, `[default: …]`) are
-            // intentionally ignored once a value-type has been established.
-            if value_type.is_none() {
-                value_type = Some(
-                    piece
-                        .trim_matches(|c| c == '<' || c == '>' || c == '[' || c == ']')
-                        .to_string(),
-                );
-            }
-        }
-    }
-    (short, long, value_type)
-}
-
-/// Collect continuation lines: lines whose indent > `base_indent` (or blank).
-/// Returns the combined description string.
-fn collect_continuation<'a>(
-    lines: &'a [&'a str],
-    start: usize,
-    base_indent: usize,
-) -> (String, usize) {
-    let mut desc_parts = Vec::new();
-    let mut i = start;
-    while i < lines.len() {
-        let line = lines[i];
-        if line.trim().is_empty() {
-            // blank line ends the block
-            break;
-        }
-        let ind = indent_of(line);
-        if ind <= base_indent {
-            break;
-        }
-        // A deeper-indented line that starts with `-` is another flag entry,
-        // not description text.  This handles tools like cargo that mix indent
-        // levels within a single Options section (e.g. indent-2 for short+long
-        // flags and indent-6 for long-only flags).
-        if line.trim().starts_with('-') {
-            break;
-        }
-        desc_parts.push(line.trim().to_string());
-        i += 1;
-    }
-    (desc_parts.join(" "), i)
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// clap parser
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Parse clap-style `--help` output.
-///
-/// Typical structure:
-/// ```text
-/// <description>
-///
-/// Usage: <name> [OPTIONS] [COMMAND]
-///
-/// Commands:
-///   subcommand1  Short description
-///   subcommand2  Short description
-///
-/// Arguments:
-///   [ARG]  description
-///
-/// Options:
-///       --flag <VALUE>
-///           Long description
-///   -h, --help
-///           Print help
-/// ```
-pub fn parse_help_clap(help: &str) -> Command {
-    let lines: Vec<&str> = help.lines().collect();
-    let mut cmd = Command::default();
-
-    // Extract name from "Usage:" line; handles both "Usage: name ..." (clap/commander)
-    // and "Usage:\n  name ..." (cobra/docopt) styles.
-    if let Some(usage_pos) = lines
-        .iter()
-        .position(|l| l.trim_start().starts_with("Usage:"))
-    {
-        let after = lines[usage_pos].trim_start_matches(|c: char| c.is_whitespace());
-        let after = after.strip_prefix("Usage:").unwrap_or(after).trim();
-        if let Some(name_part) = after.split_whitespace().next() {
-            // Standard "Usage: name ..." form.
-            cmd.name = Some(name_part.to_string());
-        } else {
-            // Cobra/docopt form: "Usage:" is alone on its own line; name is on the next.
-            for next_line in &lines[usage_pos + 1..] {
-                let t = next_line.trim();
-                if !t.is_empty() {
-                    if let Some(name_part) = t.split_whitespace().next() {
-                        cmd.name = Some(name_part.to_string());
-                    }
-                    break;
-                }
-            }
-        }
-    }
-
-    // First non-empty line before "Usage:" is the description.
-    let usage_pos = lines
-        .iter()
-        .position(|l| l.trim_start().starts_with("Usage:"))
-        .unwrap_or(0);
-    for line in &lines[..usage_pos] {
-        let t = line.trim();
-        if !t.is_empty() {
-            cmd.description = Some(t.to_string());
-            break;
-        }
-    }
-
-    // Section parsing.
-    let mut i = 0;
-    while i < lines.len() {
-        let trimmed = lines[i].trim();
-
-        // ── Commands / Subcommands section ──────────────────────────────────
-        if trimmed == "Commands:" || trimmed == "Subcommands:" || trimmed == "Available Commands:" {
-            i += 1;
-            while i < lines.len() {
-                let line = lines[i];
-                if line.trim().is_empty() {
-                    i += 1;
-                    break;
-                }
-                // Guard: a non-indented non-empty line is a new section header.
-                if indent_of(line) == 0 && !line.trim().is_empty() {
-                    break;
-                }
-                let trimmed_line = line.trim();
-                // Format: "  subname  Short description"
-                let mut parts = trimmed_line.splitn(2, "  ");
-                let sub_name = parts.next().unwrap_or("").trim().to_string();
-                let sub_desc = parts.next().map(|s| s.trim().to_string());
-                if !sub_name.is_empty() {
-                    cmd.subcommands.push(Command {
-                        name: Some(sub_name),
-                        description: sub_desc,
-                        ..Command::default()
-                    });
-                }
-                i += 1;
-            }
-            continue;
-        }
-
-        // ── Options / Arguments / Flags section ────────────────────────────
-        if trimmed == "Options:"
-            || trimmed == "Arguments:"
-            || trimmed == "Flags:"
-            || trimmed == "Global Flags:"
-            || trimmed == "Global Options:"
-            || trimmed == "Options are:"
-        {
-            i += 1;
-            while i < lines.len() {
-                let line = lines[i];
-                if line.trim().is_empty() {
-                    i += 1;
-                    // blank line may separate args but we continue
-                    continue;
-                }
-                // A non-indented non-empty line is a new section header.
-                if indent_of(line) == 0 {
-                    break;
-                }
-                let flag_indent = indent_of(line);
-                let flag_part = line.trim();
-
-                // Skip pure description continuation lines that are deeper
-                // (they belong to the previous arg).
-                // We identify flag lines as ones that start with '-' or '['.
-                if !flag_part.starts_with('-') && !flag_part.starts_with('[') {
-                    i += 1;
-                    continue;
-                }
-
-                // Only parse tokens up to the first double-space to avoid
-                // treating description text as flag tokens.
-                let token_part = flag_part
-                    .find("  ")
-                    .map(|pos| &flag_part[..pos])
-                    .unwrap_or(flag_part);
-                let (short, long, value_type) = parse_flag_tokens(token_part);
-
-                // Capture inline description (text after first double-space on
-                // the flag line), then append any deeper-indented continuation.
-                let inline_desc: Option<String> = flag_part.find("  ").and_then(|pos| {
-                    let d = flag_part[pos..].trim();
-                    if d.is_empty() {
-                        None
-                    } else {
-                        Some(d.to_string())
-                    }
-                });
-                let (cont_desc, next_i) = collect_continuation(&lines, i + 1, flag_indent);
-                i = next_i;
-
-                let description = match (inline_desc, cont_desc.as_str()) {
-                    (Some(d), "") => Some(d),
-                    (None, d) if !d.is_empty() => Some(d.to_string()),
-                    (Some(d), rest) => Some(format!("{} {}", d, rest)),
-                    (None, _) => None,
-                };
-
-                let num_args = if value_type.is_some() {
-                    Some("1".to_string())
-                } else {
-                    None
-                };
-
-                cmd.args.push(Arg {
-                    short,
-                    long,
-                    description,
-                    value_type,
-                    num_args,
-                });
-            }
-            continue;
-        }
-
-        i += 1;
-    }
-
-    cmd
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// argparse parser
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Parse Python argparse-style `--help` output.
-///
-/// Typical structure:
-/// ```text
-/// usage: script.py [-h] [--flag VALUE] positional
-///
-/// Description of the program.
-///
-/// positional arguments:
-///   positional   help text
-///
-/// optional arguments:
-///   -h, --help   show this help message and exit
-///   --flag VALUE
-///                help text for flag
-/// ```
-pub fn parse_help_argparse(help: &str) -> Command {
-    let lines: Vec<&str> = help.lines().collect();
-    let mut cmd = Command::default();
-
-    // Command name from "usage:" line.
-    if let Some(usage_line) = lines
-        .iter()
-        .find(|l| l.trim_start().to_lowercase().starts_with("usage:"))
-    {
-        let after = usage_line
-            .trim_start_matches(|c: char| c.is_whitespace())
-            .to_lowercase();
-        let after = after.strip_prefix("usage:").unwrap_or(&after).trim();
-        if let Some(name) = after.split_whitespace().next() {
-            cmd.name = Some(name.to_string());
-        }
-    }
-
-    // Description: first non-empty line after blank line following "usage:".
-    let usage_pos = lines
-        .iter()
-        .position(|l| l.trim_start().to_lowercase().starts_with("usage:"))
-        .unwrap_or(0);
-    let mut found_blank = false;
-    for line in &lines[usage_pos + 1..] {
-        if line.trim().is_empty() {
-            found_blank = true;
-        } else if found_blank {
-            let t = line.trim();
-            // Don't treat section headers as the description.
-            if !t.ends_with(':') {
-                cmd.description = Some(t.to_string());
-            }
-            break;
-        }
-    }
-
-    // Section parsing.
-    let mut i = 0;
-    while i < lines.len() {
-        let trimmed = lines[i].trim();
-
-        let is_args_section = trimmed == "positional arguments:"
-            || trimmed == "optional arguments:"
-            || trimmed == "options:"
-            || trimmed.starts_with("named arguments:");
-
-        if is_args_section {
-            i += 1;
-            while i < lines.len() {
-                let line = lines[i];
-                if line.trim().is_empty() {
-                    i += 1;
-                    break;
-                }
-                // Non-indented line → new section.
-                if indent_of(line) == 0 && line.trim().ends_with(':') {
-                    break;
-                }
-                let flag_indent = indent_of(line);
-                let flag_part = line.trim();
-
-                // argparse flag lines start with '-' for flags or a word for positional.
-                let (short, long, value_type) = if flag_part.starts_with('-') {
-                    // Only parse tokens up to the first double-space to avoid
-                    // treating description text as flag tokens.
-                    let token_part = flag_part
-                        .find("  ")
-                        .map(|pos| &flag_part[..pos])
-                        .unwrap_or(flag_part);
-                    parse_flag_tokens(token_part)
-                } else {
-                    // Positional argument: first word is the name.
-                    let name_token = flag_part.split_whitespace().next().unwrap_or("");
-                    (None, Some(name_token.to_string()), None)
-                };
-
-                // Description may be inline (after double-space) or on continuation lines.
-                let inline_desc: Option<String> = if let Some(pos) = flag_part.find("  ") {
-                    let d = flag_part[pos..].trim();
-                    if !d.is_empty() {
-                        Some(d.to_string())
-                    } else {
-                        None
-                    }
-                } else {
-                    None
-                };
-
-                let (cont_desc, next_i) = collect_continuation(&lines, i + 1, flag_indent);
-                i = next_i;
-
-                let description = match (inline_desc, cont_desc.as_str()) {
-                    (Some(d), "") => Some(d),
-                    (None, d) if !d.is_empty() => Some(d.to_string()),
-                    (Some(d), rest) => Some(format!("{} {}", d, rest)),
-                    (None, _) => None,
-                };
-
-                let num_args = if value_type.is_some() {
-                    Some("1".to_string())
-                } else {
-                    None
-                };
-
-                cmd.args.push(Arg {
-                    short,
-                    long,
-                    description,
-                    value_type,
-                    num_args,
-                });
-            }
-            continue;
-        }
-
-        i += 1;
-    }
-
-    cmd
-}
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Generic / fallback parser
-// ──────────────────────────────────────────────────────────────────────────────
-
-/// Fallback parser for unrecognised help formats.
-///
-/// Uses heuristics:
-/// - Lines that start with `-` are treated as argument definitions.
-/// - A two-space-separated suffix on such a line is the inline description.
-/// - Deeper-indented lines following a flag are appended to its description.
-pub fn parse_help_generic(help: &str) -> Command {
-    let lines: Vec<&str> = help.lines().collect();
-    let mut cmd = Command::default();
-
-    // Try to find a name from the first "usage"/"Usage" line.
-    for line in &lines {
-        let lower = line.trim().to_lowercase();
-        if lower.starts_with("usage:") {
-            let after = line
-                .trim()
-                .trim_start_matches(|c: char| c.is_alphabetic() || c == ':')
-                .trim();
-            if let Some(name) = after.split_whitespace().next() {
-                cmd.name = Some(name.to_string());
-                break;
-            }
-        }
-    }
-
-    // Collect description from the first non-empty line if it doesn't look
-    // like a flag or section header.
-    for line in &lines {
-        let t = line.trim();
-        if !t.is_empty() && !t.starts_with('-') && !t.to_lowercase().starts_with("usage:") {
-            cmd.description = Some(t.to_string());
-            break;
-        }
-    }
-
-    // Scan for argument lines.
-    let mut i = 0;
-    while i < lines.len() {
-        let line = lines[i];
-        let flag_part = line.trim();
-
-        if flag_part.starts_with('-') {
-            let flag_indent = indent_of(line);
-            // Only parse tokens up to the first double-space to avoid
-            // treating description text as flag tokens.
-            let token_part = flag_part
-                .find("  ")
-                .map(|pos| &flag_part[..pos])
-                .unwrap_or(flag_part);
-            let (short, long, value_type) = parse_flag_tokens(token_part);
-
-            let inline_desc: Option<String> = if let Some(pos) = flag_part.find("  ") {
-                let d = flag_part[pos..].trim();
-                if !d.is_empty() {
-                    Some(d.to_string())
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-            let (cont_desc, next_i) = collect_continuation(&lines, i + 1, flag_indent);
-            i = next_i;
-
-            let description = match (inline_desc, cont_desc.as_str()) {
-                (Some(d), "") => Some(d),
-                (None, d) if !d.is_empty() => Some(d.to_string()),
-                (Some(d), rest) => Some(format!("{} {}", d, rest)),
-                (None, _) => None,
-            };
-
-            let num_args = if value_type.is_some() {
-                Some("1".to_string())
-            } else {
-                None
-            };
-
-            cmd.args.push(Arg {
-                short,
-                long,
-                description,
-                value_type,
-                num_args,
-            });
-            continue;
-        }
-
-        i += 1;
-    }
-
-    cmd
 }
 
 // ──────────────────────────────────────────────────────────────────────────────
@@ -812,7 +259,7 @@ fn load_manpage_command(command_path: &str) -> anyhow::Result<Command> {
     let manpage_path = locate_manpage(&cmd_name)?;
     let manpage_content = read_manpage_source(&manpage_path)?;
 
-    man::parse_manpage(&cmd_name, &manpage_content)
+    parse_man::parse_manpage(&cmd_name, &manpage_content)
         .ok_or_else(|| anyhow::anyhow!("failed to parse man page for '{}'", cmd_name))
 }
 
@@ -2273,7 +1720,7 @@ Record intent-to-add without content.
         let manpage_loader = |command_path: &str| -> anyhow::Result<Command> {
             manpage_calls.set(manpage_calls.get() + 1);
             assert_eq!(command_path, "git-add");
-            Ok(man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
+            Ok(parse_man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
         };
 
         let cmd = synthesize_completion_with(
@@ -2331,7 +1778,7 @@ Record intent-to-add without content.
         };
         let manpage_loader = |_: &str| -> anyhow::Result<Command> {
             manpage_calls.set(manpage_calls.get() + 1);
-            Ok(man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
+            Ok(parse_man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
         };
 
         let cmd = synthesize_completion_with(
